requirements.txt
Flask==3.1.2
numpy==2.3.5
fusepy==3.0.1
blinker==1.9.0
click==8.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3
